requirements.txt
alembic==1.14.1
asgiref==3.8.1
backports.zoneinfo==0.2.1
bidict==0.23.1
blinker==1.8.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==44.0.1
defusedxml==0.7.1
Django==4.2.19
django-allauth==0.58.2
django-crispy-forms==2.0
dnspython==2.6.1
email_validator==2.2.0
Flask==3.1.0
Flask-Cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.4.1
oauthlib==3.2.2
passlib==1.7.4
Pillow==10.1.0
psycopg==3.2.5
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.0.3
render==1.0.0
requests==2.32.3
requests-oauthlib==2.0.0
simple-websocket==1.1.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.1.2
zipp==3.20.2
Flask-Limiter==3.10.1
Flask-Marshmallow==1.3.0
Flask-SeaSurf==1.1.1
Flask-Talisman==1.1.0
bcrypt==4.1.2
pytest==8.0.2
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
mypy==1.8.0
sentry-sdk==1.40.6
